requirements.txt
streamlit
langchain-openai
langchain-core
weaviate-client
duckduckgo-search
python-dotenv
pypdf
requests
